Created a separate file for simulations
