+++ src/CMakeLists.txt
@@ -23,6 +23,7 @@ add_library(libmata STATIC
 	nfa/delta.cc
 	nfa/operations.cc
 	nfa/builder.cc
+    nfa/reductionSimulation.cc
 
 	nft/nft.cc
 	nft/inclusion.cc
